browserstack test fixes
