+++ android/BS_pytest.ini
@@ -1,0 +1,6 @@
+[pytest]
+markers =
+    regression: mark a test as regression.
+    sanity: mark a test as sanity.
+
+addopts = --driver BrowserStack --capability browserName Appium --variables caps.json -v -s --html=report.html --self-contained-html --disable-pytest-warnings
