--- android/LOCAL_pytest.ini
@@ -1,6 +1,0 @@
-[pytest]
-markers =
-    regression: mark a test as regression.
-    sanity: mark a test as sanity.
-
-addopts = --driver Appium --variables caps.json -v -s --html=report.html --self-contained-html --disable-pytest-warnings
