+++ android/locators/product_detail_locators.py
@@ -4,7 +4,7 @@ from appium.webdriver.common.mobileby import MobileBy
 # LIST OF LOCATORS FOR SEARCH PAGE MODEL
 WISHLIST_STAR_BUTTON = "com.socialsuperstore.feature_product_detail:id/saveToWishlistBtn"
 ADD_BUTTON_IN_WISHLIST = "/hierarchy/android.widget.FrameLayout/android.widget.LinearLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.widget.FrameLayout/android.view.ViewGroup/android.widget.FrameLayout/android.widget.ScrollView/android.view.ViewGroup/androidx.recyclerview.widget.RecyclerView/android.view.ViewGroup/android.widget.TextView[3]"
-BACK_BTN = "Перейти вверх"
+BACK_BTN = "Navigate up"
 PRODUCT_NAME_TITLE = "com.socialsuperstore:id/toolbarText"
 PROFILE_FOOTER_ITEM = "com.socialsuperstore:id/action_profile"
 WISHLIST_PROFILE_CLICK = "com.socialsuperstore:id/profileWishlistsTitle"
